requirements.txt
Flask
Flask-JWT-Extended
psycopg2-binary
python-dotenv
Werkzeug
itsdangerous
click
csvkit
